+++ lib/services/audio_synth_stub.dart
@@ -8,3 +8,5 @@ void playWaterFlow(int chainLength) {}
 void playAmpollaGlub() {}
 
 void playVictoryFanfare() {}
+
+void playVoiceFile(String assetPath) {}